Add regulation detail page and related components
